% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cash_Flow_Summary.R
\name{Cash_Flow_Summary}
\alias{Cash_Flow_Summary}
\title{Creates a Cash Flow Summary Table}
\usage{
Cash_Flow_Summary(inv_table)
}
\arguments{
\item{inv_table}{The Inventory Table for the user.}
}
\value{
A cash flow summary table as a data frame.
}
\description{
This function creates a Cash Flow Summary Table from the Inventory Table.
}
\details{
Requires: dplyr, lubridate
}
\examples{

}
