% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create_YM_Inventory_Sell_Through.R
\name{Create_YM_Inventory_Sell_Through}
\alias{Create_YM_Inventory_Sell_Through}
\title{Creates a Year, Month Inventory Sell Through Table}
\usage{
Create_YM_Inventory_Sell_Through(ym_inv_summary)
}
\arguments{
\item{ym_inv_summary}{The Inventory Summary Table to use.}
}
\value{
ym_inv_add_summary The Year Month Inventory Sold Summary Table as Data Frame.
}
\description{
This function creates a Year, Month Inventory Sell Through Table.
}
\details{
Requires: dplyr, lubridate
}
