% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create_YM_Inventory_Sold_Summary.R
\name{Create_YM_Inventory_Sold_Summary}
\alias{Create_YM_Inventory_Sold_Summary}
\title{Creates a Year, Month Inventory Sold Summary Table}
\usage{
Create_YM_Inventory_Sold_Summary(inv_table)
}
\arguments{
\item{inv_table}{The Inventory Table to use.}
}
\value{
ym_inv_add_summary The Year Month Inventory Sold Summary Table as Data Frame.
}
\description{
This function creates a Year, Month Inventory Sold Summary from an Inventory Table.
}
\details{
Requires: dplyr, lubridate
}
