% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create_YM_Sales_Summary.R
\name{Create_YM_Sales_Summary}
\alias{Create_YM_Sales_Summary}
\title{Creates a Year, Month Sales Summary Table}
\usage{
Create_YM_Sales_Summary(inv_table)
}
\arguments{
\item{inv_table}{The Inventory Table to use.}
}
\value{
ym_sales_summary The Year Month Sales Summary Table as Data Frame.
}
\description{
This function creates a Year, Month Sales Summary from an Inventory Table.
}
\details{
Requires: dplyr, lubridate
}
