% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create_YM_Table.R
\name{Create_YM_Table}
\alias{Create_YM_Table}
\title{Creates a Year, Month Table}
\usage{
Create_YM_Table(beg_year = 2019, end_year = 2021)
}
\arguments{
\item{beg_year}{The starting year for the table.}

\item{end_year}{The ending year for the table.}
}
\value{
ym_table The Year Month Table as Data Frame.
}
\description{
This function creates a Year, Month table based on a beginning and ending year.
}
\details{
Requires: dplyr
}
