% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create_YW_Cash_Flow_Cum.R
\name{Create_YW_Cash_Flow_Cum}
\alias{Create_YW_Cash_Flow_Cum}
\title{Create Cumulative Stats for Cash Flow Table}
\usage{
Create_YW_Cash_Flow_Cum(cf_table)
}
\arguments{
\item{cf_table}{The Year, Week Cash Flow table.}
}
\value{
cf_table The Year, Week Cash Flow Table as Data Frame.
}
\description{
This function Creates Cumulative Stats for Cash Flow Table.
}
\details{
Requires: dplyr, lubridate
}
