% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create_YW_Cash_Flow_In.R
\name{Create_YW_Cash_Flow_In}
\alias{Create_YW_Cash_Flow_In}
\title{Creates a Year, Week Cash Flow In Table}
\usage{
Create_YW_Cash_Flow_In(inv_table)
}
\arguments{
\item{inv_table}{The Inventory Table to use.}
}
\value{
yw_cash_flow_in The Year, Week Cash Flow In Table as Data Frame.
}
\description{
This function Creates a Year, Week Cash Flow In Table.
}
\details{
Requires: dplyr, lubridate
}
