% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create_YW_Cash_Flow_Out.R
\name{Create_YW_Cash_Flow_Out}
\alias{Create_YW_Cash_Flow_Out}
\title{Creates a Year, Week Cash Flow Out Table}
\usage{
Create_YW_Cash_Flow_Out(inv_table)
}
\arguments{
\item{inv_table}{The Inventory Table to use.}
}
\value{
yw_cash_flow_out The Year, Week Cash Flow Out Table as Data Frame.
}
\description{
This function Creates a Year, Week Cash Flow Out Table.
}
\details{
Requires: dplyr, lubridate
}
