% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create_YW_Table.R
\name{Create_YW_Table}
\alias{Create_YW_Table}
\title{Creates a Year, Week Table}
\usage{
Create_YW_Table(beg_year = 2019, end_year = 2021)
}
\arguments{
\item{beg_year}{The starting year for the table.}

\item{end_year}{The ending year for the table.}
}
\value{
yw_table The Year Month Table as Data Frame.
}
\description{
This function creates a Year, Week table based on a beginning and ending year.
}
\details{
Requires: dplyr
}
