% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Daily_Sales_Summary.R
\name{Daily_Sales_Summary}
\alias{Daily_Sales_Summary}
\title{Creates a Daily Sales Summary Table}
\usage{
Daily_Sales_Summary(inv_table)
}
\arguments{
\item{inv_table}{The Inventory Table for the user.}
}
\value{
A Daily Sales Summary Table as a data frame.
}
\description{
This function creates a Daily Sales Summary Table from the Inventory Table.
}
\details{
Requires: dplyr
}
\examples{

}
