% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inv_Bin_Summary.R
\name{Inv_Bin_Summary}
\alias{Inv_Bin_Summary}
\title{Creates a Summary Inventory Bin Table}
\usage{
Inv_Bin_Summary(b_inv_table)
}
\arguments{
\item{inv_table}{The Inventory Table for the user.}
}
\value{
A formatted table detailing inventory bin usage.
}
\description{
This function creates a Summary Inventory Bin Table and formats it using the Formattable package. This function calls Inv_Bin_Table.
}
\details{
Requires: dplyr, formattable
}
\examples{

}
