% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inv_Bin_Table.R
\name{Inv_Bin_Table}
\alias{Inv_Bin_Table}
\title{Creates an Inventory Bin Table}
\usage{
Inv_Bin_Table(inv_table)
}
\arguments{
\item{inv_table}{The Inventory Table for the user.}
}
\value{
A data frame with inventory items by bin location.
}
\description{
This function creates an Inventory Bin Table.
}
\details{
Requires: dplyr
}
\examples{

}
