% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inventory_Summary_Listed.R
\name{Inventory_Summary_Listed}
\alias{Inventory_Summary_Listed}
\title{Creates a Listed Inventory Summary Table}
\usage{
Inventory_Summary_Listed(inv_table)
}
\arguments{
\item{inv_table}{The Inventory Table for the user.}
}
\value{
A Listed Inventory Summary Table as a data frame.
}
\description{
This function creates a Listed Inventory Summary Table from the Inventory Table.
}
\details{
Requires: dplyr
}
