% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inventory_Summary_Open.R
\name{Inventory_Summary_Open}
\alias{Inventory_Summary_Open}
\title{Creates an Open Inventory Summary Table}
\usage{
Inventory_Summary_Open(inv_table)
}
\arguments{
\item{inv_table}{The Inventory Table for the user.}
}
\value{
An Open Inventory Summary Table as a data frame.
}
\description{
This function creates a Open Inventory Summary Table from the Inventory Table.
}
\details{
Requires: dplyr
}
