% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inventory_Summary_Sold.R
\name{Inventory_Summary_Sold}
\alias{Inventory_Summary_Sold}
\title{Creates a Sold Inventory Summary Table}
\usage{
Inventory_Summary_Sold(inv_table)
}
\arguments{
\item{inv_table}{The Inventory Table for the user.}
}
\value{
A Sold Inventory Summary Table as a data frame.
}
\description{
This function creates a Sold Inventory Summary Table from the Inventory Table.
}
\details{
Requires: dplyr
}
