% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inventory_Summary_Total.R
\name{Inventory_Summary_Total}
\alias{Inventory_Summary_Total}
\title{Creates a Total Inventory Summary Table}
\usage{
Inventory_Summary_Total(inv_table)
}
\arguments{
\item{inv_table}{The Inventory Table for the user.}
}
\value{
A Total Inventory Summary Table as a data frame.
}
\description{
This function creates a Total Inventory Summary Table from the Inventory Table.
}
\details{
Requires: dplyr, lubridate
}
