% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Listed_Inventory_Summary.R
\name{Listed_Inventory_Summary}
\alias{Listed_Inventory_Summary}
\title{Creates a Listed Inventory Summary Table}
\usage{
Listed_Inventory_Summary(inv_table)
}
\arguments{
\item{inv_table}{The Inventory Table for the user.}
}
\value{
A Listed Inventory Summary Table as a data frame.
}
\description{
This function creates a Listed Inventory Summary Table from the Inventory Table.
}
\details{
Requires: dplyr
}
\examples{

}
