% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Merch_Sales_Summary.R
\name{Merch_Sales_Summary}
\alias{Merch_Sales_Summary}
\title{Creates the Merchant Sales Summary Table}
\usage{
Merch_Sales_Summary(inv_table)
}
\arguments{
\item{inv_table}{The Inventory Table for the user.}
}
\value{
A Merchant Sales Summary Table as a data frame.
}
\description{
This function creates a Merchant Sales Summary Table from the Inventory Table.
}
\details{
Requires: dplyr
}
