% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Monthly_Inventory_Detail.R
\name{Monthly_Inventory_Detail}
\alias{Monthly_Inventory_Detail}
\title{Creates a Monthly Inventory Detail Table}
\usage{
Monthly_Inventory_Detail(inv_table)
}
\arguments{
\item{inv_table}{The Inventory Table for the user.}
}
\value{
A Monthly Inventory Detail Table as a data frame.
}
\description{
This function creates a Monthly Inventory Detail Table from the Inventory Table.
}
\details{
Requires: dplyr, lubridate
}
\examples{

}
