% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Monthly_Inventory_Sell_Through.R
\name{Monthly_Inventory_Sell_Through}
\alias{Monthly_Inventory_Sell_Through}
\title{Creates a Monthly Inventory Sell Through Table}
\usage{
Monthly_Inventory_Sell_Through(inv_table)
}
\arguments{
\item{inv_table}{The Inventory Table for the user.}
}
\value{
A Monthly Inventory Summary Table as a data frame.
}
\description{
This function creates a Monthly Inventory Sell Through Table from the Inventory Table.
}
\details{
Requires: dplyr, lubridate
}
