% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Monthly_Inventory_Summary.R
\name{Monthly_Inventory_Summary}
\alias{Monthly_Inventory_Summary}
\title{Creates a Monthly Inventory Summary Table}
\usage{
Monthly_Inventory_Summary(inv_table)
}
\arguments{
\item{inv_table}{The Inventory Table for the user.}
}
\value{
A Monthly Inventory Summary Table as a data frame.
}
\description{
This function creates a Monthly Inventory Summary Table from the Inventory Table.
}
\details{
Requires: dplyr, lubridate, stats
}
\examples{

}
