% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Monthly_Sales_Detail.R
\name{Monthly_Sales_Detail}
\alias{Monthly_Sales_Detail}
\title{Creates a Monthly Sales Detail Table}
\usage{
Monthly_Sales_Detail(inv_table)
}
\arguments{
\item{inv_table}{The Inventory Table for the user.}
}
\value{
A Monthly Sales Detail Table as a data frame.
}
\description{
This function creates a Monthly Sales Detail Table from the Inventory Table.
}
\details{
Requires: dplyr
}
\examples{

}
