% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Monthly_Sales_Summary.R
\name{Monthly_Sales_Summary}
\alias{Monthly_Sales_Summary}
\title{Creates a Monthly Sales Summary Table}
\usage{
Monthly_Sales_Summary(inv_table)
}
\arguments{
\item{inv_table}{The Inventory Table for the user.}
}
\value{
ym_sales_summary A Monthly Sales Summary Table as a data frame.
}
\description{
This function creates a Monthly Sales Summary Table from the Inventory Table.
}
\details{
Requires: dplyr
}
