% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Open_Inventory_Summary.R
\name{Open_Inventory_Summary}
\alias{Open_Inventory_Summary}
\title{Creates an Open Inventory Summary Table}
\usage{
Open_Inventory_Summary(inv_table)
}
\arguments{
\item{inv_table}{The Inventory Table for the user.}
}
\value{
An Open Inventory Summary Table as a data frame.
}
\description{
This function creates a Open Inventory Summary Table from the Inventory Table.
}
\details{
Requires: dplyr
}
\examples{

}
