% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PM_Sales_Monthly.R
\name{PM_Sales_Monthly}
\alias{PM_Sales_Monthly}
\title{Creates a Poshmark Sales Monthly Table}
\usage{
PM_Sales_Monthly(pm_sales)
}
\arguments{
\item{pm_sales}{The Poshmark Sales Table for the user.}
}
\value{
A Poshmark Sales Monthly Table as a data frame.
}
\description{
This function creates a Poshmark Sales Monthly Table from the Inventory Table.
}
\details{
Requires: dplyr, scales, lubridate
}
\examples{

}
