% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PM_Sales_Summary.R
\name{PM_Sales_Summary}
\alias{PM_Sales_Summary}
\title{Creates a Poshmark Sales Summary Table}
\usage{
PM_Sales_Summary(pm_sales)
}
\arguments{
\item{pm_sales}{The Poshmark Sales Table for the user.}
}
\value{
A Poshmark Sales Summary Table as a data frame.
}
\description{
This function creates a Poshmark Sales Summary Table from the Inventory Table.
}
\details{
Requires: dplyr, scales
}
\examples{

}
