% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PM_Summary_For_Chart.R
\name{PM_Summary_For_Chart}
\alias{PM_Summary_For_Chart}
\title{Creates a Summary Table for Charting}
\usage{
PM_Summary_For_Chart(t_summary, var_name)
}
\arguments{
\item{t_summary}{The summary table.}

\item{var_name}{The variable name for primary sorting.}
}
\value{
A summary table as a data frame, top 10.
}
\description{
This function Creates a Summary Table for Charting.
}
\details{
Requires: dplyr
}
\examples{

}
