% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PM_Summary_Stats_Conf.R
\name{PM_Summary_Stats_Conf}
\alias{PM_Summary_Stats_Conf}
\title{Creates Poshmark Summary Stats for File Upload Confirmation}
\usage{
PM_Summary_Stats_Conf(sales_activity)
}
\arguments{
\item{sales_activity}{Poshmark Sales Activity table.}
}
\value{
summary_stats_con Poshmark summary statistics.
}
\description{
This function Creates Poshmark Summary Stats for File Upload Confirmation.
}
\details{
Requires: dplyr
}
\examples{

}
