% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sold_Inventory_Summary.R
\name{Sold_Inventory_Summary}
\alias{Sold_Inventory_Summary}
\title{Creates a Sold Inventory Summary Table}
\usage{
Sold_Inventory_Summary(inv_table)
}
\arguments{
\item{inv_table}{The Inventory Table for the user.}
}
\value{
A Sold Inventory Summary Table as a data frame.
}
\description{
This function creates a Sold Inventory Summary Table from the Inventory Table.
}
\details{
Requires: dplyr
}
\examples{

}
