% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_Inventory_Summary.R
\name{Source_Inventory_Summary}
\alias{Source_Inventory_Summary}
\title{Creates a Source Inventory Summary Table}
\usage{
Source_Inventory_Summary(inv_table)
}
\arguments{
\item{inv_table}{The Inventory Table for the user.}
}
\value{
A Source Inventory Summary Table as a data frame.
}
\description{
This function creates a Source Inventory Summary Table from the Inventory Table.
}
\details{
Requires: dplyr
}
\examples{

}
