% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Total_Inventory_Summary.R
\name{Total_Inventory_Summary}
\alias{Total_Inventory_Summary}
\title{Creates a Total Inventory Summary Table}
\usage{
Total_Inventory_Summary(inv_table)
}
\arguments{
\item{inv_table}{The Inventory Table for the user.}
}
\value{
A Total Inventory Summary Table as a data frame.
}
\description{
This function creates a Total Inventory Summary Table from the Inventory Table.
}
\details{
Requires: dplyr, lubridate
}
\examples{

}
